requirements.txt
praw
openai
python-telegram-bot
python-dotenv
